move get_current_user_info to db utils
